Fits CMP_mu discrete kernel smoother for a sample of counts

This function computes the density using the CMP_mu discrete kernel smooth over a grid of points using the given bandwidth h or dispersion parameter nu.

compak_evalpmf(a.sample, x = NULL, h = NULL, nu = NULL, workers = 1L)

Arguments

a.sample
numeric vector: the data sample from which estimate is to be computed.

x
Either NULL or a numeric vector: the points of the grid at which the density is to be estimated. If none is provided, the range of a.sample will be used.

h, nu
numeric: the bandwidth or smoothing parameter. Only one is needed and they are related by nu = 1/h.

workers
numeric; a positive integer to represent the number of cores used for parallel processing to evaluate the kde

Value

A list containing the following components:

f.cmp

the estimated p.m.f. values

kernel.est

a list that contains the estimated kernel at each grid point

Examples

data(days)
# The bandwidth h can be the one obtained by cross validation.
(h.CV <- compak_CVbandwidth(days))

#> [1] 0.0250636

compak_evalpmf(days, 20:40, h = h.CV, workers = 1)

#> $f.cmp
#>  [1] 4.194719e-11 9.547839e-08 3.406180e-05 2.089986e-03 2.493491e-02
#>  [6] 7.535728e-02 1.069462e-01 1.342435e-01 1.608490e-01 1.505088e-01
#> [11] 8.083472e-02 4.716466e-02 6.163341e-02 6.476709e-02 4.808763e-02
#> [16] 2.930260e-02 1.124434e-02 1.880767e-03 1.181614e-04 2.705051e-06
#> [21] 2.274496e-08
#> 
#> $kernel.est
#> $kernel.est[[1]]
#>  [1] 2.138844e-09 4.864513e-06 1.729079e-03 1.043142e-01 1.151892e+00
#>  [6] 2.495337e+00 1.130425e+00 1.136060e-01 2.675429e-03 1.553580e-05
#> [11] 2.332587e-08 9.466079e-12 1.082357e-15 3.625506e-20 3.690488e-25
#> [16] 1.181712e-30 1.229711e-36 4.288775e-43 5.161359e-50 2.203431e-57
#> [21] 3.425592e-65
#> 
#> $kernel.est[[2]]
#>  [1] 4.624457e-13 4.883110e-09 8.058375e-06 2.257105e-03 1.157166e-01
#>  [6] 1.163827e+00 2.447804e+00 1.142120e+00 1.248761e-01 3.366628e-03
#> [11] 2.346794e-05 4.421633e-08 2.347245e-11 3.650332e-15 1.725135e-19
#> [16] 2.564640e-24 1.239063e-29 2.006314e-35 1.120999e-41 2.221859e-48
#> [21] 1.603718e-55
#> 
#> $kernel.est[[3]]
#>  [1] 4.362572e-17 2.020436e-12 1.462389e-08 1.796528e-05 4.039659e-03
#>  [6] 1.781986e-01 1.643839e+00 3.364041e+00 1.613227e+00 1.907560e-01
#> [11] 5.832098e-03 4.819470e-05 1.122127e-07 7.653896e-11 1.586500e-14
#> [16] 1.034452e-18 2.192017e-23 1.556741e-28 3.814957e-34 3.316404e-40
#> [21] 1.049895e-46
#> 
#> $kernel.est[[4]]
#>  [1] 1.138013e-21 2.192676e-16 6.602626e-12 3.374522e-08 3.156804e-05
#>  [6] 5.793378e-03 2.223368e-01 1.892944e+00 3.776563e+00 1.857821e+00
#> [11] 2.363062e-01 8.124080e-03 7.869396e-05 2.233092e-07 1.925700e-10
#> [16] 5.223761e-14 4.605127e-18 1.360626e-22 1.387192e-27 5.016939e-33
#> [21] 6.607572e-39
#> 
#> $kernel.est[[5]]
#>  [1] 1.303647e-26 9.948394e-21 1.186482e-15 2.401723e-11 8.898653e-08
#>  [6] 6.468066e-05 9.831504e-03 3.315221e-01 2.619614e+00 5.103997e+00
#> [11] 2.571270e+00 3.501166e-01 1.343216e-02 1.509653e-04 5.156145e-07
#> [16] 5.539695e-10 1.934239e-13 2.263462e-17 9.139806e-22 1.309198e-26
#> [21] 6.829271e-32
#> 
#> $kernel.est[[6]]
#>  [1] 7.709065e-33 2.225495e-26 1.004077e-20 7.688831e-16 1.077689e-11
#>  [6] 2.963300e-08 1.703935e-05 2.173589e-03 6.497319e-02 4.788937e-01
#> [11] 9.126584e-01 4.701162e-01 6.822920e-02 2.900904e-03 3.748118e-05
#> [16] 1.523372e-07 2.012158e-10 8.907523e-14 1.360669e-17 7.373146e-22
#> [21] 1.454969e-26
#> 
#> $kernel.est[[7]]
#>  [1] 5.200382e-39 5.440579e-32 8.895496e-26 2.468587e-20 1.253911e-15
#>  [6] 1.249492e-11 2.603729e-08 1.203663e-05 1.303907e-03 3.482864e-02
#> [11] 2.405418e-01 4.490269e-01 2.361687e-01 3.638900e-02 1.703864e-03
#> [16] 2.509646e-05 1.201305e-07 1.927229e-10 1.066876e-13 2.095072e-17
#> [21] 1.498251e-21
#> 
#> $kernel.est[[8]]
#>  [1] 1.230053e-44 4.479572e-37 2.549555e-30 2.462893e-24 4.354784e-19
#>  [6] 1.510555e-14 1.095725e-10 1.763249e-07 6.649029e-05 6.182312e-03
#> [11] 1.486304e-01 9.658117e-01 1.768257e+00 9.484099e-01 1.545837e-01
#> [16] 7.925818e-03 1.320651e-04 7.375147e-07 1.421196e-09 9.714983e-13
#> [21] 2.418414e-16
#> 
#> $kernel.est[[9]]
#>  [1] 3.351444e-51 4.090970e-43 7.804331e-36 2.526960e-29 1.497618e-23
#>  [6] 1.741214e-18 4.233493e-14 2.283455e-10 2.886147e-07 8.994825e-05
#> [11] 7.248216e-03 1.578692e-01 9.687944e-01 1.741662e+00 9.515094e-01
#> [16] 1.635215e-01 9.132724e-03 1.709482e-04 1.104154e-06 2.529874e-09
#> [21] 2.110907e-12
#> 
#> $kernel.est[[10]]
#>  [1] 2.201900e-58 8.693969e-50 5.364803e-42 5.618789e-35 1.077138e-28
#>  [6] 4.050877e-23 3.185825e-18 5.558298e-14 2.272449e-10 2.290839e-07
#> [11] 5.971166e-05 4.206798e-03 8.350495e-02 4.855911e-01 8.581163e-01
#> [16] 4.770174e-01 8.617591e-02 5.217667e-03 1.090103e-04 8.079105e-07
#> [21] 2.180518e-09
#> 
#> $kernel.est[[11]]
#>  [1] 1.454759e-65 1.796659e-56 3.467813e-48 1.136054e-40 6.812112e-34
#>  [6] 8.013340e-28 1.971245e-22 1.075758e-17 1.375692e-13 4.337864e-10
#> [11] 3.536672e-07 7.793659e-05 4.839006e-03 8.801748e-02 4.865176e-01
#> [16] 8.459426e-01 4.780208e-01 9.052978e-02 5.916118e-03 1.371471e-04
#> [21] 1.157810e-06
#> 
#>